refactor: Shop 타입 구조 이동
