+++ src/hooks/api/user/useGetMyInfoQuery.ts
@@ -1,4 +1,4 @@
-import { User } from "@/types/user";
+import { User } from "@/types/api/user";
 import { instance } from "@/utils/instance";
 import { useQuery } from "@tanstack/react-query";
