+++ src/hooks/api/user/useLoginQuery.ts
@@ -1,6 +1,6 @@
 import { instance } from "@/utils/instance";
 import { useMutation } from "@tanstack/react-query";
-import { LoginRequest, LoginResponse } from "@/types/user";
+import { LoginRequest, LoginResponse } from "@/types/api/user";
 
 const postLogin = async ({ email, password }: LoginRequest): Promise<LoginResponse> => {
   const response = await instance.post("/token", { email, password });
