+++ src/types/api/shop.ts
@@ -1,0 +1,10 @@
+export interface Shop {
+  id: string;
+  name: string;
+  category: string;
+  address1: string;
+  address2: string;
+  description: string;
+  imageUrl: string;
+  originalHourlyPay: number;
+}
